+++ module/Rest/src/Action/ShortUrl/ListShortUrlsAction.php
@@ -4,14 +4,12 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Rest\Action\ShortUrl;
 
-use Cake\Chronos\Chronos;
-use InvalidArgumentException;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Log\LoggerInterface;
 use Shlinkio\Shlink\Common\Paginator\Util\PaginatorUtilsTrait;
-use Shlinkio\Shlink\Common\Util\DateRange;
+use Shlinkio\Shlink\Core\Model\ShortUrlsParams;
 use Shlinkio\Shlink\Core\Service\ShortUrlServiceInterface;
 use Shlinkio\Shlink\Core\Transformer\ShortUrlDataTransformer;
 use Shlinkio\Shlink\Rest\Action\AbstractRestAction;
@@ -36,38 +34,11 @@ class ListShortUrlsAction extends AbstractRestAction
         $this->domainConfig = $domainConfig;
     }
 
-    /**
-     * @throws InvalidArgumentException
-     */
     public function handle(Request $request): Response
     {
-        $params = $this->queryToListParams($request->getQueryParams());
-        $shortUrls = $this->shortUrlService->listShortUrls(...$params);
+        $shortUrls = $this->shortUrlService->listShortUrls(ShortUrlsParams::fromRawData($request->getQueryParams()));
         return new JsonResponse(['shortUrls' => $this->serializePaginator($shortUrls, new ShortUrlDataTransformer(
             $this->domainConfig,
         ))]);
-    }
-
-    /**
-     * @param array $query
-     * @return array
-     */
-    private function queryToListParams(array $query): array
-    {
-        return [
-            (int) ($query['page'] ?? 1),
-            $query['searchTerm'] ?? null,
-            $query['tags'] ?? [],
-            $query['orderBy'] ?? null,
-            $this->determineDateRangeFromQuery($query),
-        ];
-    }
-
-    private function determineDateRangeFromQuery(array $query): DateRange
-    {
-        return new DateRange(
-            isset($query['startDate']) ? Chronos::parse($query['startDate']) : null,
-            isset($query['endDate']) ? Chronos::parse($query['endDate']) : null,
-        );
     }
 }